% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_tf_get.R
\name{browse_gotf_main}
\alias{browse_gotf_main}
\title{use DT::datatable to browse the GO catalogue of human DNA-binding transcription factors in Table S1.A of Lovering et al.}
\usage{
browse_gotf_main(cache = BiocFileCache::BiocFileCache(ask = FALSE))
}
\arguments{
\item{cache}{a BiocFileCache instance}
}
\value{
result of DT::datatable
}
\description{
use DT::datatable to browse the GO catalogue of human DNA-binding transcription factors in Table S1.A of Lovering et al.
}
\examples{
if (interactive()) browse_gotf_main()
}
