% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambert_ccbr_get.R
\name{browse_humantfs_main}
\alias{browse_humantfs_main}
\title{use DT::datatable to browse the Lambert's Human Transcription Factors repository}
\usage{
browse_humantfs_main(cache = BiocFileCache::BiocFileCache(ask = FALSE))
}
\arguments{
\item{cache}{a BiocFileCache instance}
}
\value{
result of DT::datatable
}
\description{
use DT::datatable to browse the Lambert's Human Transcription Factors repository
}
\examples{
if (interactive()) browse_humantfs_main()
}
