% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_lambert_gwaslinks.R
\name{browse_lambert_gwaslinks}
\alias{browse_lambert_gwaslinks}
\title{browse several hundred disease-TF associations with hyperlinked PMIDs}
\usage{
browse_lambert_gwaslinks()
}
\value{
DT::datatable
}
\description{
browse several hundred disease-TF associations with hyperlinked PMIDs
}
\note{
Based on supplemental table S4 of PMID 29425488
}
\examples{
if (interactive()) browse_lambert_gwaslinks()
}
