% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cisbpTFcat}
\alias{cisbpTFcat}
\title{cisbpTFcat: data.frame with information on CISBP TFs for human, retained for reproducibility support; see
cisbpTFcat_2.0 for a more recent catalog}
\format{
data.frame
}
\source{
\url{http://cisbp.ccbr.utoronto.ca/bulk.php} select Homo_sapiens
}
\usage{
cisbpTFcat
}
\description{
cisbpTFcat: data.frame with information on CISBP TFs for human, retained for reproducibility support; see
cisbpTFcat_2.0 for a more recent catalog
}
\note{
Extracted March 2018, checked August 2018.  The only changes
observed are that genes ZUFSP and T are used has HGNC values
in the March catalog; these symbols seem to be absent from the
org.Hs.eg.db of August 2018.  The records involved are
1356, 7412 and 7413.  These symbols were left in the package image
of CISBP in August 2018.
}
\examples{
head(TFutils::cisbpTFcat)
}
\keyword{datasets}
