% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cisbpTFcat_2.0}
\alias{cisbpTFcat_2.0}
\title{cisbpTFcat_2.0: data.frame with information on CISBP TFs for human, described in PMID 31133749}
\format{
data.frame
}
\source{
\url{http://cisbp.ccbr.utoronto.ca/bulk.php} select Homo_sapiens
}
\usage{
cisbpTFcat_2.0
}
\description{
cisbpTFcat_2.0: data.frame with information on CISBP TFs for human, described in PMID 31133749
}
\note{
Extracted August 2019.
}
\examples{
head(TFutils::cisbpTFcat_2.0)
}
\keyword{datasets}
