% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfcirc.R
\name{defaultCircosParms}
\alias{defaultCircosParms}
\title{basic layout parameters for circos}
\usage{
defaultCircosParms()
}
\value{
a list
}
\description{
basic layout parameters for circos
}
\examples{
head(defaultCircosParms())
}
