% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepi.R
\name{directHitsInCISBP}
\alias{directHitsInCISBP}
\title{demonstrate interoperation of TF catalog with GWAS catalog}
\usage{
directHitsInCISBP(traitTag, gwascat)
}
\arguments{
\item{traitTag}{character(1) string found in DISEASE/TRAIT field of gwascat instance}

\item{gwascat}{instance of \code{\link[gwascat]{gwaswloc-class}}}
}
\value{
data.frame
}
\description{
demonstrate interoperation of TF catalog with GWAS catalog
}
\examples{
data(gwascat_hg19_chr17)
directHitsInCISBP("Prostate cancer" , gwascat_hg19_chr17)
}
