% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfutils.R
\name{getBS}
\alias{getBS}
\title{acquire a GRanges representing binding sites for a TF}
\usage{
getBS(conv_name, map, which, meth)
}
\arguments{
\item{conv_name}{character(1) conventional name for TF}

\item{map}{list with elements consisting of the notation for TFs; names of list elements are the conv_names}

\item{which}{a GRanges instance limiting the request}

\item{meth}{a function with two parameters, tf (a character(1) Mnnnn
string) and which (a GRanges instance) that retrieves TF 'bed'-like
records from a resource}
}
\description{
acquire a GRanges representing binding sites for a TF
}
\note{
We assume that the retrieval yields a data.frame in the format\preformatted{    V1    V2    V3               V4        V5 V6       V7 
 chr1 11496 11511 chr1:11496-11511 -1.555560  - 5.92e-04 
 chr1 11520 11535 chr1:11520-11535 -3.161620  + 9.12e-04 
}
}
\examples{
\dontrun{
data(named_tf)
requireNamespace("GenomicRanges")
requireNamespace("IRanges")
getBS("VDR", named_tf, GenomicRanges::GRanges("chr1", IRanges::IRanges(1,25000)), 
     importFIMO_local)
}
}
