% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwascat_GRCh38}
\alias{gwascat_GRCh38}
\title{gwascat_GRCh38: GRanges of march 21 2018 EBI gwascat}
\format{GenomicRanges GRanges instance}
\source{
gwascat::makeCurrentGwascat, with a little extra metadata
}
\usage{
gwascat_GRCh38
}
\description{
gwascat_GRCh38: GRanges of march 21 2018 EBI gwascat
}
\examples{
TFutils::gwascat_GRCh38[,1:5]
}
\keyword{datasets}
