% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwascat_hg19}
\alias{gwascat_hg19}
\title{gwascat_hg19: GRanges of march 21 2018 EBI gwascat}
\format{GenomicRanges GRanges instance}
\source{
gwascat::makeCurrentGwascat, with gwascat:::lo38to19 applied
}
\usage{
gwascat_hg19
}
\description{
gwascat_hg19: GRanges of march 21 2018 EBI gwascat
}
\examples{
TFutils::gwascat_hg19[,1:5]
}
\keyword{datasets}
