% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hocomoco.mono.sep2018}
\alias{hocomoco.mono.sep2018}
\title{hocomoco.mono.sep2018: data.frame with information on HOCOMOCO TFs for human, Sept 2018 download}
\format{data.frame}
\source{
\url{http://hocomoco11.autosome.ru/human/mono?full=true}
}
\usage{
hocomoco.mono.sep2018
}
\description{
hocomoco.mono.sep2018: data.frame with information on HOCOMOCO TFs for human, Sept 2018 download
}
\note{
Extracted September 2018
}
\examples{
head(TFutils::hocomoco.mono.sep2018)
}
\keyword{datasets}
