% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{importFIMO_S3}
\alias{importFIMO_S3}
\title{utility to read FIMO outputs (after sorting and tabix indexing applied as in contents of biocfound-tfbs S3 bucket)}
\usage{
importFIMO_S3(tf, which)
}
\arguments{
\item{tf}{character(1) file id}

\item{which}{a GRanges delimiting the extract}
}
\description{
utility to read FIMO outputs (after sorting and tabix indexing applied as in contents of biocfound-tfbs S3 bucket)
}
\examples{
requireNamespace("GenomicRanges")
requireNamespace("IRanges")
importFIMO_S3("M5946_1", 
    which=GenomicRanges::GRanges("chr1", IRanges::IRanges(1,15000)))
}
