% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_file.R
\name{include_file}
\alias{include_file}
\title{include contents of a file}
\usage{
include_file(file, ..., encoding = "UTF-8")
}
\arguments{
\item{file}{file path, length 1}

\item{...}{other file component}

\item{encoding}{which encoding does the file use? defaults to UTF-8}
}
\description{
include contents of a file in a target file, e.g., a Rmd file.
}
