% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lambert_snps}
\alias{lambert_snps}
\title{lambert_snps is Table S3 of Lambert et al PMID 29425488}
\format{
data.frame
}
\usage{
lambert_snps
}
\description{
lambert_snps is Table S3 of Lambert et al PMID 29425488
}
\examples{
head(lambert_snps)
}
\keyword{datasets}
