% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfutils.R
\name{mapTF}
\alias{mapTF}
\title{utility to allow user to retrieve the map of tf name (symbol or motif_id)}
\usage{
mapTF(tf)
}
\arguments{
\item{tf}{character string with either the motif_id or the hgnc_symbol}
}
\description{
utility to allow user to retrieve the map of tf name (symbol or motif_id)
}
\examples{
data(motif_id_map)
data(symbol_map)
mapTF("VDR")
}
