% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{metadata_tf}
\alias{metadata_tf}
\title{metadata_tf: metadata (motif_if and hgnc_symbol) about all the CISBP FIMO scan TF bed files}
\format{list}
\usage{
metadata_tf
}
\description{
metadata_tf: metadata (motif_if and hgnc_symbol) about all the CISBP FIMO scan TF bed files
}
\examples{
data(metadata_tf)
}
\keyword{datasets}
