% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seqinfo_hg19_chr17}
\alias{seqinfo_hg19_chr17}
\title{a Seqinfo instance for a chr17 in hg19}
\format{
a Seqinfo instance
}
\usage{
seqinfo_hg19_chr17
}
\description{
a Seqinfo instance for a chr17 in hg19
}
\examples{
seqinfo_hg19_chr17
}
\keyword{datasets}
