% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupHIZE.R
\name{setupHIZE}
\alias{setupHIZE}
\title{process a gene_attribute_matrix.txt file from harmonizeome into a GeneSetCollection}
\usage{
setupHIZE(txtfn = "gene_attribute_matrix.txt", tag)
}
\arguments{
\item{txtfn}{character(1) path to gene_attribute_matrix.txt file from harmonizeome}

\item{tag}{character(1) will be added to shortDescription field of each GeneSet instance}
}
\value{
GSEABase::GeneSetCollection
}
\description{
process a gene_attribute_matrix.txt file from harmonizeome into a GeneSetCollection
}
\note{
After uncompressing content of \url{http://amp.pharm.mssm.edu/static/hdfs/harmonizome/data/cheappi/gene_attribute_matrix.txt.gz} run this on gene_attribute_matrix.txt with tag="CHEA".
}
