% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{tfbs_as_GRanges}
\alias{tfbs_as_GRanges}
\title{acquire a GRanges representing binding sites for a TF}
\usage{
tfbs_as_GRanges(conv_name, map, which, meth, genomeTag)
}
\arguments{
\item{conv_name}{character(1) conventional name for TF}

\item{map}{data.frame with columns PWMid (in form Mnnnn_n,
denoting the PWM defining the motif used for FIMO scoring) and
HGNC (the conventional name for the TF, as a 'gene symbol')}

\item{which}{a GRanges instance limiting the request}

\item{meth}{a function with two parameters, tf (a character(1) Mnnnn
string) and which (a GRanges instance) that retrieves TF 'bed'-like
records from a resource}

\item{genomeTag}{character(1) tag for genome build in use}
}
\description{
acquire a GRanges representing binding sites for a TF
}
\note{
We assume that the retrieval yields a data.frame in the format\preformatted{    V1    V2    V3               V4        V5 V6       V7 
 chr1 11496 11511 chr1:11496-11511 -1.555560  - 5.92e-04 
 chr1 11520 11535 chr1:11520-11535 -3.161620  + 9.12e-04
}

As of 21 Mar 2018 there are only two TFs in S3: VDR, and
}
\examples{
requireNamespace("GenomicRanges")
requireNamespace("IRanges")
tfbs_as_GRanges(conv_name="VDR", map=TFutils::fimoMap, 
     which=GenomicRanges::GRanges("chr1", IRanges::IRanges(1,25000), 
     genomeTag="hg19"), meth=importFIMO_S3)
}
