% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfcirc.R
\name{tffamCirc.plot}
\alias{tffamCirc.plot}
\title{use a radial plot (by default) for motif stack}
\usage{
tffamCirc.plot(motiflist, circosParms = defaultCircosParms())
}
\arguments{
\item{motiflist}{a list of pfm instances from motifStack}

\item{circosParms}{a list of parameter settings for circos plot}
}
\value{
side effect to graphics device
}
\description{
use a radial plot (by default) for motif stack
}
\examples{
p1 = tffamCirc.prep( )
tffamCirc.plot(p1[c(1:8, 10:17, 19)])
}
