% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfcirc.R
\name{tffamCirc.prep}
\alias{tffamCirc.prep}
\title{set up list of pfms in motifStack protocol}
\usage{
tffamCirc.prep(tffam = "Paired-related HD factors{3.1.3}",
  trimfac = 0.4)
}
\arguments{
\item{tffam}{character(1) name of TF family as found in TFutils::hocomoco.mono field \code{TF family}}

\item{trimfac}{fraction passed as parameter \code{t} to motifStack::trimMotif}
}
\value{
a list of pfm instances as defined in motifStack
}
\description{
set up list of pfms in motifStack protocol
}
\note{
Uses MotifDb, motifStack to create a list of pfms
}
\examples{
n1 = tffamCirc.prep()
str(n1)
}
