% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tfhash}
\alias{tfhash}
\title{tfhash: data.frame with MSigDb TFs, TF targets as symbol or ENTREZ}
\format{list}
\usage{
tfhash
}
\description{
tfhash: data.frame with MSigDb TFs, TF targets as symbol or ENTREZ
}
\examples{
TFutils::tfhash
tfhash[1:3,]
}
\keyword{datasets}
