% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gob.R
\name{TFtargs}
\alias{TFtargs}
\title{gadget to help sort through tags naming TFs}
\usage{
TFtargs(gscoll = TFutils::tftColl, initTF = "VDR")
}
\arguments{
\item{gscoll}{a GSEABase GeneSetCollection}

\item{initTF}{character(1) initial TF string for app}
}
\description{
gadget to help sort through tags naming TFs
}
\note{
Will use gwascat_hg19 to look for 'MAPPED_GENE' field entries matching targets.
}
