% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tftColl}
\alias{tftColl}
\title{tftColl: GSEABase GeneSetCollection for transcription factor targets}
\format{GSEABase GeneSetCollection instance}
\source{
broad institute
}
\usage{
tftColl
}
\description{
tftColl: GSEABase GeneSetCollection for transcription factor targets
}
\examples{
data(tftColl)
tftColl
}
\keyword{datasets}
