% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfutils.R
\name{importFIMO_local}
\alias{importFIMO_local}
\title{utility to read FIMO outputs from local resource(cluster)}
\usage{
importFIMO_local(tf, which)
}
\arguments{
\item{tf}{character(1) file id}

\item{which}{a GRanges delimiting the extract}
}
\description{
utility to read FIMO outputs from local resource(cluster)
}
\examples{
requireNamespace("GenomicRanges")
requireNamespace("IRanges")
requireNamespace("DT")
importFIMO_local("M5946", param=GenomicRanges::GRanges("chr1", IRanges::IRanges(1,15000)))
}
