% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfutils.R, R/tfutils2.R
\name{importFIMO_local}
\alias{importFIMO_local}
\alias{importFIMO_local}
\title{utility to read FIMO outputs from local resource(cluster)}
\usage{
importFIMO_local(tf, which)

importFIMO_local(tf, which)
}
\arguments{
\item{tf}{character(1) file id}

\item{which}{a GRanges delimiting the extract}

\item{tf}{character(1) file id}

\item{which}{a GRanges delimiting the extract}
}
\description{
utility to read FIMO outputs from local resource(cluster)

utility to read FIMO outputs from local resource(cluster)
}
\examples{
\dontrun{  # need data local in the package
requireNamespace("GenomicRanges")
requireNamespace("IRanges")
requireNamespace("DT")
importFIMO_local("M5946_1", which=GenomicRanges::GRanges("chr1", IRanges::IRanges(1,15000)))
} 
requireNamespace("GenomicRanges")
requireNamespace("IRanges")
importFIMO_local("M5946_1", which=GenomicRanges::GRanges("chr1", IRanges::IRanges(1,15000)))
}
