% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_tf_get.R
\name{retrieve_gotf_main}
\alias{retrieve_gotf_main}
\title{acquire the content of Table S1.A from Lovering et al., A GO catalogue of human DNA-binding
transcription factors, DOI: https://doi.org/10.1101/2020.10.28.359232}
\usage{
retrieve_gotf_main(cache = BiocFileCache::BiocFileCache(ask = FALSE))
}
\arguments{
\item{cache}{a BiocFileCache instance}
}
\value{
a tbl_df
}
\description{
acquire the content of Table S1.A from Lovering et al., A GO catalogue of human DNA-binding
transcription factors, DOI: https://doi.org/10.1101/2020.10.28.359232
}
\note{
This will download the spreadsheet if not found in \code{cache}.
}
\examples{
if (interactive()) retrieve_gotf_main()
}
