% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambert_ccbr_get.R
\name{retrieve_humantfs_main}
\alias{retrieve_humantfs_main}
\title{acquire the CSV content for table S1 of Lambert et al. Cell 2018, "The Human Transcription Factors" from the Human TFS website}
\usage{
retrieve_humantfs_main(cache = BiocFileCache::BiocFileCache(ask = FALSE))
}
\arguments{
\item{cache}{a BiocFileCache instance}
}
\value{
a tbl_df
}
\description{
acquire the CSV content for table S1 of Lambert et al. Cell 2018, "The Human Transcription Factors" from the Human TFS website
}
\note{
This will download the spreadsheet if not found in \code{cache}.
}
\examples{
if (interactive()) retrieve_lambert_main()
}
