% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFCatalog-class.R
\docType{class}
\name{TFCatalog-class}
\alias{TFCatalog-class}
\alias{"TFCatalog-class"}
\title{define a structure to hold information about TFs from diverse reference sources}
\description{
define a structure to hold information about TFs from diverse reference sources
}
